requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pyodbc==5.1.0
python-dotenv==1.0.1
pydantic==2.6.4
